hng64: Hacked 3d acknowledge in a slightly different manner.  Sams64 games now show 3d. [Andrew Gardner]
NOTES NOT IN WHATSNEW:
This makes all games in hng64.c display a good amount of 3d.
The read handler I modified definitely seems to have something to do with
the missing 3d in these games.  It's only read if the frame being drawn
is missing geometry.  My new hack is a lucky guess, but hopefully we'll
figure out what this function is really looking for sometime soon.

Snaps here: http://www.mameworld.info/ajg/