snes.c: simplified drawing routines, added support for offset-per-tile, cleaned up the code [Fabio Priuli]
out of whatsnew: again, this is the first of a series of patches that I separately commit to make easier to trace back regressions. the whole set of changes can be mentioned in next release with the above subject.

For people interested in details, these are the changes in this step 1:

* changed pixel drawing to only draw pixels 0-256 in both non-hires and hires modes. as a result hires graphics is temporarily broken (BGs squeezed and only half drawn)