Remove the following functions:
  cpu_get_total_cycles()   == cpudevice->total_cycles()
  cpu_clocks_to_attotime() == cpudevice->cycles_to_attotime()
  cpu_attotime_to_clocks() == cpudevice->attotime_to_cycles()

In some cases, I updated existing code that fetched and cached 
devices to fetch them as cpu_device so that these can be called
without casting.

In other cases, I did the quick & dirty thing which was to downcast
a generic device to a cpu_device and then call the function.

Note that cpu_clocks_to_attotime() and cpu_attotime_to_clocks()
were previously mapping to the device-level functions
clocks_to_attotime() and attotime_to_clocks(). This was different
behavior than before the big devices change, so with this update
I have fixed it to match the previous behavior (i.e., map cycles
not clocks).

This may impact timing of drivers that used these functions on
CPUs that have clock dividers or multipliers.