Small fix to separate shared inputs for glfgreat (tmnt.c)
Comment:
It's hard to know if a 4th button is even used as the test mode only shows SHOT1 through SHOT3.  Pressing SHOT4 for player one as mapped toggles the all 3 bits that are to the right of the DIP bank status display at the bottom of the test screen.  Pressing mapped buttons for player 2-4 do not appear to have any effect.