SDL: Support -uifont for using arbitrary TrueType fonts on Linux/BSD builds.
This takes full pathname/filename to the font, and the name can have styles 
appended as in Windows: [b] for bold, [i] for italic, [s] for strikethrough, 
and [u] for underlined.  Fonts are rendered at a size of 120 points and scaled 
down by the core, which looks superb on high-res displays.  (Liberation Sans
that comes with most recent distros looks really good).

Warning: Linux/BSD builds now require SDL_ttf 2.x.  On Fedora the required 
package names are "SDL_ttf" and "SDL_ttf-devel".