New games marked as GAME_NOT_WORKING ------------------------------------ Arcana Heart Full [Tormod, Smitdogg, The Dumping Union]
Comments from Haze: Note, I haven't marked this as a bootleg, but 75% of the rom is actually MAME ROM zips, they've clearly recycled flash chips from one of the MAME multi-boots, so I have a feeling that this might have been a 3rd party upgrade.  I doubt they shipped like this from the factory, so I've added a comment to that effect because it should definitely be verified against other copies of the game at some point.