new NOT WORKING Grasspin [Alberto Grego, Zabanitu]
off what's new, it seems another Zilic / Ashby developed game, Christoper Stamper is credited in the sound rom.

marked as NOT working because it always gives me infinite lives? dips are read by the sound cpu and passed to the main cpu, but coinage / # lives works, so i'm not sure what's going on? help?