idectrl: words 57/58 of IDENTIFY DEVICE are total sectors, not sectors per track [R. Belmont]
As commented, this is the case for ATA-1, -2, -3, -4, and -5.  The CFFA2 Apple 
II IDE card firmware relies on this value being as documented in the standards 
to properly detect the drive size, and none of the usual sensitive Atari/Midway
games are broken by the change.

(This field is unused in ATA-6 and later).